// src/app/apis/model/module/private/modulo.ts
export class Modulo {
    codigo: number;
    nombreModulo: string;
    descripcion: string;
    subtitulo: string;
    icono: string;
    codigoAplicacion: string;
    descripcionAplicacion: string;
    estadoRegistro: string;

    constructor() {
        this.codigo = 0;
        this.nombreModulo = "";
        this.descripcion = "";
        this.subtitulo = "";
        this.icono = "";
        this.codigoAplicacion = "";
        this.descripcionAplicacion = "";
        this.estadoRegistro = "";
    }
}